OLH-1698: Add browser monitor for webchat widget
We are introducing a JS widget on the contact page that will load a
webchat client. This will give users another route to get support from
One Login.

The webchat is a service hosted by one of our suppliers and we have
limited monitoring available to us.

Add a new browser monitor that loads the English and Welsh contact pages
and waits for the 'start webchat' button to appear. This button is
injected by the JS, so once it appears we know that our supplier's
origin is up and that the webchat has successfully loaded enough for a
user to open the chat.

The English and Welsh chats are deployed as two separate applications by
our supplier, so it's necessary to load both languages.

The monitor is configured to check against our integration in the
non-production Dynatrace and production in the production Dynatrace.

I've set `enabled = false` for the production Dynatrace for now as we're
not expecting to launch to production until at least next week. The chat
is up and running in integration so that's enabled.