Fix per-minute graphs for sign-ups and sign-ins service-health-overview-tsd.json
There is an issue with the current metric expressions for these two graphs. When you zoom out the timeframe Dynatrace will force a time resolution of greater than 1 min. With the current metric expression this will aggregate the values instead of displaying them as a rate of per-minute.

This fixes the metric expression so that the graphs work correctly and resolve to a per-minute value regardless of the timeframe selected.

Signed-off-by: Surfraz Ahmed <surfraz@users.noreply.github.com>